remove dcTools > dc
